Moving the structure generators to generator.py
